Do not SPAM pull_replication_log when updating drafts
The update of draft comments relies on the continuous forced
update of the refs/drafts-comments/<> in the All-Users project.
Performing a forced update with the apply-object is rejected
and the operation should naturally fallback to a git fetch.

There is no need to display a full stacktrace of the exception
and report the REJECTED execution status on the pull_replication_log.

Change-Id: I566e16043a482686d455d16ac7ac865d1d8352d7