Consume only relevant stream events
The pull-replication plugin is uninterested in the vast majority of
ProjectEvent specifications, which include numerous types such as
ChangeAbandoned, CommentAdded, and FetchRefReplicatedEvent, among
others.

Instead, pull-replication focuses solely on events that trigger the
creation, update, or removal of Git repositories. Specifically, these
events are:

- ProjectDeletedEvent
- ProjectCreatedEvent
- RefUpdatedEvent
- ProjectHeadUpdatedEvent

All other ProjectEvent types are unnecessary for pull-replication and
can be safely ignored when consuming stream events. This change ensures
that only relevant events are processed, improving efficiency and
reducing noise.

Change-Id: I5914f3752d44776eb6e2c25ec9b7ca60a0e6eeee