Report deletions as succeeded in the fetch-ref-replicated events
The refs deletions were always reported as failing on the
stream event fetch-ref-replicated, because of a mismatch in the
ref-spec reported for deletions.

Prefix the refs deletions with ":", as per git specs[1], for reporting them
correctly in the status map and therefore generating more
useful events on the replication status.

[1] https://git-scm.com/book/en/v2/Git-Internals-The-Refspec under "Deleting References"

Change-Id: Ibd1b9c97cdc79f4972a65cba8001530bf0a64ab5