Ban synchronous create ref via ref-update in apply-object(s)
Ref creation is a non-fast-forward operation and it may
have different outcome with other forced operations when
not executed in the correct order.
Without a precedessor, it is impossible for the apply-object
to understand if the replication can be applied or not
and it may risk to apply the updates out of order.

For example, when replicating a series of create/delete
of the same ref, the order could be inverted because
of the create executed synchronously and the delete
asynchronously, generating a different final state
compared to the source replication endpoint.

Ban the use of ref-update via apply-object returning a 400
and force the client to run an asynchronous fetch instead
which would also go through global-refdb validation.

This is similar to the approach implemented with
I978c3079fbb9.

NOTE: For refs/changes/* it is correct to assume that
all ref-updates are associated with immutable refs because
even though changes can be deleted, their change numbers
cannot be reused. For refs/changes/*/meta the check for
its predecessor is already implemented and works as expected.

Bug: Issue 397737226
Change-Id: I486c00570bc728f38dbe8ffc781f5b05bc66cd76