Move ProjectDeletionAction/CacheDeleteHandler PullReplicationModule
Following I31cb7378ba7a the classes responsible for providing the delete
project functionality need to be moved to PullReplicationModule as it is
used also in the stream event listener and therefore used even without
the HTTP REST-APIs.

Having the binding in the HttpModule caused two conflicting bindings,
one done just-in-time indirectly by the StreamEventListener and another
one explicitly bound in the PullReplicationModule which was making the
plugin failing to load because of a double binding.

Change-Id: I9dd29693ee48851a45bff79078f7de27d7c631fc