Allow locking refs whilst filtering for fetch requests
Allow the lock/unlock pattern when filtering the refs to fetch. This
mechanism is used for making sure that the refs selected for the fetch
replication operation do not move throughout the replication process.

This is needed for preventing a ref-update operation happening
concurrently with the replication process invalidating the filtering
validity.

Change-Id: I6a30aff43a81f00d671554186eb6e796446b148d