Do not notify failures on collisions
Currently, when a replication task is rescheduled due to a collision,
it is not marked as a retry.

However, the failure notification logic only suppresses notifications
for tasks explicitly flagged as retries.

This change adds an additional check to prevent failure notifications
for tasks that are rescheduled because of collisions.

Bug: Issue 418811956
Change-Id: If489f6b9ac4cdb307a4332bbf23dcf2b51f5178a