Consider the headName value when creating a new repository
The creation of a new repository is notified through the ProjectCreatedEvent
but the headName was ignored, causing the new project to be always created
with a default head.

Use the headName in the ProjectCreatedEvent and use it for creating
the new empty repository.

Also consider the 'headName' parameter when the creation happens via REST-API,
even though the client-side of the replication via HTTP isn't implemented
yet in this change and will be expanded in a follow-up change.

Bug: Issue 430486435
Change-Id: I2a37b01e55c1acbc5e51bde365dd84c9943c9b31