Prevent concurrent fetches on the same repository
When multiple replication tasks attempt to fetch from the
same repository at the same time, each may lock a different ref
and then wait indefinitely for the other’s lock to be released.
This might lead to deadlocks.

To avoid this, ensure that only one fetch is executed at a time
per repository.

Bug: Issue 437805590
Change-Id: I7883e680e018705fcc0d307356b1c541fe8ecc93